.ci/pylint_requirements.txt
-r ../requirements.txt
pylint-sqlalchemy>=0.2.0
